revise: adds triplet id with version and patch to reference genome
Adds the concept of a uniquely identifying triplet for reference
genomes. With the triplet, you should immediately be able to identify
(1) what the colloquial name of the reference genome is, (2) if the
reference genome is versioned (outside of the name itself), what version
is it based off of, and (3) what is the base genome upon which this
reference genome is built.