revise: better parsing of bioinformatics file formats
Previously, we were doing ad-hoc parsing of bioinformatics file formats
all across the codebase by looking at file extensions. This was really
messy, and meant that there were many places that needed to be updated
to support a new file format. Now, all of that is unified in the
`formats` module.