revise: puts some reference genomes behind feature flag
Reference genomes are a big portion of the compile time for `ngs`. In
many cases, this extensive list of reference genomes is not needed
unless you are actually using the command itself. Thus, I have gated the
majority of the reference genomes behind the
`extended-reference-genomes` feature flag. This should reduce compile
times while also keeping the feature available when publishing the tool.